+++ src/Casting/Sexadecimal/Round.php
@@ -3,6 +3,7 @@ namespace MarcoConsiglio\Goniometry\Casting\Sexadecimal;
 
 use MarcoConsiglio\Goniometry\Casting\Sexagesimal;
 use MarcoConsiglio\Goniometry\Interfaces\Casting\ToSexadecimal;
+use MarcoConsiglio\Goniometry\Interfaces\SexadecimalValue;
 use MarcoConsiglio\Goniometry\SexadecimalDegrees;
 use RoundingMode;
 
@@ -15,7 +16,7 @@ class Round extends Sexagesimal implements ToSexadecimal
      * Construct the `Round` object.
      */
     public function __construct(
-        protected SexadecimalDegrees $sexadecimal,
+        protected SexadecimalValue $sexadecimal,
         protected int|null $precision = null
     ) {}