+++ tests/Feature/AngularDistanceTest.php
@@ -1,0 +1,56 @@
+<?php
+namespace MarcoConsiglio\Goniometry\Tests\Feature;
+
+use MarcoConsiglio\Goniometry\AngularDistance;
+use MarcoConsiglio\Goniometry\Builders\AngularDistance\FromSexadecimal;
+use MarcoConsiglio\Goniometry\Casting\Sexadecimal\Round;
+use MarcoConsiglio\Goniometry\Casting\Sexagesimal;
+use MarcoConsiglio\Goniometry\Degrees;
+use MarcoConsiglio\Goniometry\Minutes;
+use MarcoConsiglio\Goniometry\Random\Generator\PositiveSexadecimal as PositiveSexadecimalGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\RelativeSexadecimal as RelativeSexadecimalGenerator;
+use MarcoConsiglio\Goniometry\Random\Validator\FloatValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\PositiveSexadecimal as PositiveSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\RelativeSexadecimal as RelativeSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Seconds;
+use MarcoConsiglio\Goniometry\SexadecimalAngularDistance;
+use MarcoConsiglio\Goniometry\SexagesimalDegrees;
+use MarcoConsiglio\Goniometry\Tests\TestCase;
+use MarcoConsiglio\Goniometry\Traits\WithAngleFaker;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\TestDox;
+use PHPUnit\Framework\Attributes\UsesClass;
+use PHPUnit\Framework\Attributes\UsesTrait;
+
+#[CoversClass(AngularDistance::class)]
+#[UsesClass(FromSexadecimal::class)]
+#[UsesClass(Round::class)]
+#[UsesClass(Sexagesimal::class)]
+#[UsesClass(Degrees::class)]
+#[UsesClass(Minutes::class)]
+#[UsesClass(PositiveSexadecimalGenerator::class)]
+#[UsesClass(RelativeSexadecimalGenerator::class)]
+#[UsesClass(FloatValidator::class)]
+#[UsesClass(PositiveSexadecimalValidator::class)]
+#[UsesClass(RelativeSexadecimalValidator::class)]
+#[UsesClass(Seconds::class)]
+#[UsesClass(SexadecimalAngularDistance::class)]
+#[UsesClass(SexagesimalDegrees::class)]
+#[UsesTrait(WithAngleFaker::class)]
+class AngularDistanceTest extends TestCase
+{
+    #[TestDox("can be created from a decimal number.")]
+    public function test_create_from_decimal(): void
+    {
+        // Arrange
+        $decimal = $this->randomSexadecimal(precision: 1);
+
+        // Act
+        $angle = AngularDistance::createFromDecimal($decimal);
+
+        $this->assertEquals(
+            $decimal, 
+            $angle->toFloat(1)
+        );
+    }
+}