+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+## Unreleased
+## Added
+- `AngularDistance` class to represent angular distance between two objects.
+- `AngularDistance::createFromDecimal()` method to build an `AngularDistance` object from a `float` or `SexadecimalAngularDistance` type input.
+- `SexadecimalAngularDistance` class to represent sexadecimal value of an `AngularDistance` object.
+- `AngularDistance::toSexadecimalAngularDistance()` to cast an `AngularDistance` object to a `SexadecimalAngularDistance` object.
+- `AngularDistance::toFloat()` method to cast an `AngularDistance` object to sexadecimal `float` value.
+
 ## v4.1.1 2025-04-29
 ### Changed
 - API documentation.